+++ RxGRDB/SelectStatementRequest+Rx.swift
@@ -45,7 +45,7 @@ extension Reactive where Base: SelectStatementRequest {
     public func fetchCount(
         in writer: DatabaseWriter,
         startImmediately: Bool = true,
-        scheduler: ImmediateSchedulerType = MainScheduler.instance)
+        scheduler: ImmediateSchedulerType? = nil)
         -> Observable<Int>
     {
         let request = base